requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
Automat==22.10.0
bokeh==3.3.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.1
databricks-sdk==0.11.0
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
gradio==3.47.1
gradio_client==0.6.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
hyperlink==21.0.0
idna==3.4
importlib-resources==6.1.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
orjson==3.9.9
packaging==23.2
pandas==2.1.1
parsel==1.8.1
Pillow==10.0.1
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pydub==0.25.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rpds-py==0.10.6
Scrapy==2.11.0
semantic-version==2.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tldextract==5.0.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
w3lib==2.1.2
websockets==11.0.3
xyzservices==2023.10.0
zope.interface==6.1
